refactor(e2e): decompose models.py into focused sub-modules
Split src/scylla/e2e/models.py (931 LOC) into focused modules under
src/scylla/e2e/models_internals/. Public API preserved via thin facade
re-exporting every previously public name. Move-only refactor; no
behavior changes.

Refs #1876
Refs #1867
