+++ src/scylla/e2e/models_internals/__init__.py
@@ -1,0 +1,5 @@
+"""Internal sub-modules for :mod:`scylla.e2e.models`.
+
+These modules are implementation details — public callers should continue to
+import everything from :mod:`scylla.e2e.models`.
+"""
